backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
pandas==2.2.2
python-jose==3.3.0
supabase>=2.10.0
httpx>=0.25.2
requests==2.31.0
# ...
python-dotenv==1.0.1
fpdf2>=2.7.0
sdv>=0.2.12
scikit-learn>=1.5.0
# PyTorch ecosystem (pinned for Mac Apple Silicon compatibility)
# Using torch 2.2.2 (compatible with SynthCity <2.3 requirement)
# Opacus 1.5.4 is patched to work with torch 2.2.2 (rms_norm.py handles missing RMSNorm)
torch==2.2.2
opacus==1.5.4


scipy<1.13

# OMOP Engine
sentence-transformers==2.2.2
huggingface-hub==0.19.4
faiss-cpu==1.7.4
# pandas>=2.0.0 (Already satisfied by pandas==2.2.2 above)
sdmetrics>=0.14.0
